+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 ### New features
 
 * Adding `Norb/MisplacedBranchingLogic` cop to ensure branching logic is not in the wrong place. ([@tleish][])
+* Adding `Norb/NamespacedActiveRecord` cop to ensure ActiveRecord models are namespaced. ([@tleish][])
 
 
 [@tleish]: https://github.com/tleish